+++ hutool-extra/src/test/java/cn/hutool/extra/pinyin/PinyinUtilTest.java
@@ -1,5 +1,6 @@
 package cn.hutool.extra.pinyin;
 
+import cn.hutool.extra.pinyin.engine.pinyin4j.Pinyin4jEngine;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -8,6 +9,13 @@ public class PinyinUtilTest {
 	@Test
 	public void getPinyinTest(){
 		final String pinyin = PinyinUtil.getPinyin("你好", " ");
+		Assert.assertEquals("ni hao", pinyin);
+	}
+
+	@Test
+	public void getPinyinByPinyin4jTest(){
+		final Pinyin4jEngine engine = new Pinyin4jEngine();
+		final String pinyin = engine.getPinyin("你好", " ");
 		Assert.assertEquals("ni hao", pinyin);
 	}
 
@@ -21,5 +29,12 @@ public class PinyinUtilTest {
 	public void getFirstLetterTest(){
 		final String result = PinyinUtil.getFirstLetter("H是第一个", ", ");
 		Assert.assertEquals("h, s, d, y, g", result);
+	}
+
+	@Test
+	public void getFirstLetterByPinyin4jTest(){
+		final Pinyin4jEngine engine = new Pinyin4jEngine();
+		final String result = engine.getFirstLetter("林海", "");
+		Assert.assertEquals("lh", result);
 	}
 }